[JIT] Unshare types for modules that define() in __init__
**Summary**
By default, scripting tries to share concrete and JIT types across
compilations. However, this can lead to incorrect results if a module
extends `torch.jit.ScriptModule`, and injects instance variables into
methods defined using `define`.

This commit detects when this has happened and disables type sharing
for the compilation of the module that uses `define` in `__init__`.

**Test Plan**
This commit adds a test to TestTypeSharing that tests this scenario.

**Fixes**
This commit fixes #43580.

ghstack-source-id: 90e6b36527fb7a3ce950b5a2a05d09027cd32c97
Pull Request resolved: https://github.com/pytorch/pytorch/pull/44233